refactor(api): extract import preflight diagnostics
