+++ app/controllers/api/v1/imports_controller.rb
@@ -3,11 +3,6 @@
 class Api::V1::ImportsController < Api::V1::BaseController
   include Pagy::Backend
 
-  PREFLIGHT_CSV_IMPORT_TYPES = %w[
-    TransactionImport TradeImport AccountImport MintImport CategoryImport RuleImport
-  ].freeze
-  PREFLIGHT_IMPORT_TYPES = (PREFLIGHT_CSV_IMPORT_TYPES + [ "SureImport" ]).freeze
-
   # Ensure proper scope authorization
   before_action :ensure_read_scope, only: [ :index, :show, :preflight ]
   before_action :ensure_write_scope, only: [ :create ]
@@ -125,15 +120,8 @@ class Api::V1::ImportsController < Api::V1::BaseController
   end
 
   def preflight
-    family = current_resource_owner.family
-    type = preflight_import_type
-    return unless type
-
-    if type == "SureImport"
-      preflight_sure_import
-    else
-      preflight_csv_import(family, type)
-    end
+    response = Import::Preflight.new(family: current_resource_owner.family, params: params).call
+    render json: response.payload, status: response.status
   rescue ActiveRecord::RecordNotFound
     render json: {
       error: "record_not_found",
@@ -193,310 +181,6 @@ class Api::V1::ImportsController < Api::V1::BaseController
         :amount_type_strategy,
         :amount_type_inflow_value
       )
-    end
-
-    def preflight_import_type
-      type = params[:type].to_s
-      return "TransactionImport" if type.blank?
-      return type if PREFLIGHT_IMPORT_TYPES.include?(type)
-
-      render json: {
-        error: "invalid_import_type",
-        message: "type must be one of: #{PREFLIGHT_IMPORT_TYPES.join(', ')}",
-        errors: [ "type must be one of: #{PREFLIGHT_IMPORT_TYPES.join(', ')}" ]
-      }, status: :unprocessable_entity
-      nil
-    end
-
-    def preflight_sure_import
-      content, filename, content_type = preflight_sure_import_upload_attributes
-      return unless content
-
-      render json: {
-        data: sure_import_preflight_payload(content, filename, content_type)
-      }
-    end
-
-    def preflight_csv_import(family, type)
-      content, filename, content_type = preflight_csv_upload_attributes
-      return unless content
-
-      import = family.imports.build(import_config_params.merge(type: type, raw_file_str: content))
-      import.account = family.accounts.find(params[:account_id]) if params[:account_id].present?
-      apply_preflight_import_defaults(import)
-
-      unless import.requires_csv_workflow?
-        render json: {
-          error: "unsupported_import_type",
-          message: "Preflight supports CSV import types and SureImport."
-        }, status: :unprocessable_entity
-        return
-      end
-
-      import.valid?
-      csv_content = preflight_csv_content(import, content)
-      csv = Import.parse_csv_str(csv_content, col_sep: import.col_sep)
-      row_count = csv.length
-      csv_headers = Array(csv.headers).compact
-      missing_required_headers = preflight_missing_required_headers(import, csv_headers)
-      errors = import.errors.full_messages.map do |message|
-        { code: "validation_failed", message: message }
-      end
-
-      if missing_required_headers.any?
-        errors << {
-          code: "missing_required_headers",
-          message: "Missing required columns: #{missing_required_headers.join(', ')}"
-        }
-      end
-
-      if row_count.zero?
-        errors << {
-          code: "no_data_rows",
-          message: "No data rows were found."
-        }
-      end
-
-      warnings = []
-      warnings << "Row count exceeds this import type's publish limit." if row_count > import.max_row_count
-
-      render json: {
-        data: {
-          type: type,
-          valid: errors.empty?,
-          content: preflight_content_payload(filename, content_type, content),
-          stats: {
-            rows_count: row_count,
-            valid_rows_count: row_count,
-            invalid_rows_count: 0
-          },
-          headers: csv_headers,
-          required_headers: preflight_required_header_labels(import),
-          missing_required_headers: missing_required_headers,
-          errors: errors,
-          warnings: warnings
-        }
-      }
-    end
-
-    def preflight_csv_upload_attributes
-      if params[:file].present?
-        preflight_csv_file_upload_attributes(params[:file])
-      elsif params[:raw_file_content].present?
-        preflight_csv_raw_content_attributes(params[:raw_file_content].to_s)
-      else
-        render json: {
-          error: "missing_content",
-          message: "Provide a CSV file or raw_file_content."
-        }, status: :unprocessable_entity
-        nil
-      end
-    end
-
-    def preflight_csv_file_upload_attributes(file)
-      if file.size > Import::MAX_CSV_SIZE
-        render json: {
-          error: "file_too_large",
-          message: "File is too large. Maximum size is #{Import::MAX_CSV_SIZE / 1.megabyte}MB."
-        }, status: :unprocessable_entity
-        return
-      end
-
-      unless Import::ALLOWED_CSV_MIME_TYPES.include?(file.content_type)
-        render json: {
-          error: "invalid_file_type",
-          message: "Invalid file type. Please upload a CSV file."
-        }, status: :unprocessable_entity
-        return
-      end
-
-      [
-        file.read,
-        file.original_filename.presence || "import.csv",
-        file.content_type.presence || "text/csv"
-      ]
-    end
-
-    def preflight_csv_raw_content_attributes(content)
-      if content.bytesize > Import::MAX_CSV_SIZE
-        render json: {
-          error: "content_too_large",
-          message: "Content is too large. Maximum size is #{Import::MAX_CSV_SIZE / 1.megabyte}MB."
-        }, status: :unprocessable_entity
-        return
-      end
-
-      [ content, "import.csv", "text/csv" ]
-    end
-
-    def preflight_sure_import_upload_attributes
-      if params[:file].present?
-        preflight_sure_import_file_upload_attributes(params[:file])
-      elsif params[:raw_file_content].present?
-        preflight_sure_import_raw_content_attributes(params[:raw_file_content].to_s)
-      else
-        render json: {
-          error: "missing_content",
-          message: "Provide a Sure NDJSON file or raw_file_content."
-        }, status: :unprocessable_entity
-        nil
-      end
-    end
-
-    def preflight_sure_import_file_upload_attributes(file)
-      if file.size > SureImport::MAX_NDJSON_SIZE
-        render json: {
-          error: "file_too_large",
-          message: "File is too large. Maximum size is #{SureImport::MAX_NDJSON_SIZE / 1.megabyte}MB."
-        }, status: :unprocessable_entity
-        return
-      end
-
-      extension = File.extname(file.original_filename.to_s).downcase
-      unless SureImport::ALLOWED_NDJSON_CONTENT_TYPES.include?(file.content_type) || extension.in?(%w[.ndjson .json])
-        render json: {
-          error: "invalid_file_type",
-          message: "Invalid file type. Please upload a Sure NDJSON file."
-        }, status: :unprocessable_entity
-        return
-      end
-
-      [
-        file.read,
-        file.original_filename.presence || "sure-import.ndjson",
-        file.content_type.presence || "application/x-ndjson"
-      ]
-    end
-
-    def preflight_sure_import_raw_content_attributes(content)
-      if content.bytesize > SureImport::MAX_NDJSON_SIZE
-        render json: {
-          error: "content_too_large",
-          message: "Content is too large. Maximum size is #{SureImport::MAX_NDJSON_SIZE / 1.megabyte}MB."
-        }, status: :unprocessable_entity
-        return
-      end
-
-      [ content, "sure-import.ndjson", "application/x-ndjson" ]
-    end
-
-    def sure_import_preflight_payload(content, filename, content_type)
-      line_counts = Hash.new(0)
-      errors = []
-      valid_rows_count = 0
-      nonblank_rows_count = 0
-
-      content.each_line.with_index(1) do |line, line_number|
-        next if line.strip.blank?
-
-        nonblank_rows_count += 1
-        record = JSON.parse(line)
-
-        unless record.is_a?(Hash)
-          errors << {
-            code: "invalid_ndjson_record",
-            message: "Line #{line_number} must be a JSON object."
-          }
-          next
-        end
-
-        if record["type"].blank? || !record.key?("data")
-          errors << {
-            code: "invalid_ndjson_record",
-            message: "Line #{line_number} must include type and data."
-          }
-          next
-        end
-
-        valid_rows_count += 1
-        line_counts[record["type"]] += 1
-      rescue JSON::ParserError => e
-        errors << {
-          code: "invalid_json",
-          message: "Line #{line_number} is not valid JSON: #{e.message}"
-        }
-      end
-
-      if nonblank_rows_count.zero?
-        errors << {
-          code: "no_data_rows",
-          message: "No data rows were found."
-        }
-      end
-
-      entity_counts = SureImport.dry_run_totals_from_ndjson(content)
-      unsupported_types = line_counts.keys - %w[
-        Account Category Tag Merchant Transaction Trade Valuation Budget BudgetCategory Rule
-      ]
-      warnings = []
-      warnings << "No importable records were found." if nonblank_rows_count.positive? && entity_counts.values.sum.zero?
-      warnings << "Some records use unsupported types: #{unsupported_types.join(', ')}" if unsupported_types.any?
-      warnings << "Row count exceeds this import type's publish limit." if nonblank_rows_count > SureImport.new.max_row_count
-
-      {
-        type: "SureImport",
-        valid: errors.empty? && nonblank_rows_count.positive?,
-        content: preflight_content_payload(filename, content_type, content),
-        stats: {
-          rows_count: nonblank_rows_count,
-          valid_rows_count: valid_rows_count,
-          invalid_rows_count: nonblank_rows_count - valid_rows_count,
-          entity_counts: entity_counts,
-          record_type_counts: line_counts
-        },
-        errors: errors,
-        warnings: warnings
-      }
-    end
-
-    def preflight_content_payload(filename, content_type, content)
-      {
-        filename: filename,
-        content_type: content_type,
-        byte_size: content.bytesize
-      }
-    end
-
-    def preflight_csv_content(import, content)
-      return content unless import.rows_to_skip.to_i.positive?
-
-      content.lines.drop(import.rows_to_skip.to_i).join
-    end
-
-    def apply_preflight_import_defaults(import)
-      return unless import.is_a?(MintImport)
-
-      import.assign_attributes(
-        signage_convention: "inflows_positive",
-        date_col_label: "Date",
-        date_format: "%m/%d/%Y",
-        name_col_label: "Description",
-        amount_col_label: "Amount",
-        currency_col_label: "Currency",
-        account_col_label: "Account Name",
-        category_col_label: "Category",
-        tags_col_label: "Labels",
-        notes_col_label: "Notes",
-        entity_type_col_label: "Transaction Type"
-      )
-    end
-
-    def preflight_required_header_labels(import)
-      import.required_column_keys.filter_map do |key|
-        import.respond_to?("#{key}_col_label") ? import.public_send("#{key}_col_label").presence || key.to_s : key.to_s
-      end
-    end
-
-    def preflight_missing_required_headers(import, headers)
-      normalized_headers = Array(headers).compact.to_h { |header| [ preflight_normalized_header(header), header ] }
-
-      preflight_required_header_labels(import).reject do |header|
-        normalized_headers.key?(preflight_normalized_header(header))
-      end
-    end
-
-    def preflight_normalized_header(header)
-      header.to_s.strip.downcase.gsub(/\*/, "").gsub(/[\s-]+/, "_")
     end
 
     def create_sure_import(family)
